♻️ REFACTOR: input 컴포넌트 삭제
