+++ src/components/SearchBar/index.js
@@ -1,4 +1,4 @@
-import { IconArea, InputTransparent } from "@components/Input/InputType";
+import IconArea from "@components/IconArea/IconArea";
 import "./SearchBar.css";
 
 export const SearchBar = () => {
@@ -6,7 +6,7 @@ export const SearchBar = () => {
     <>
       <div className="search-bar">
         <IconArea backgroundImageUrl={`/assets/images/icons/SearchIcon.png`} />
-        <InputTransparent placeholder={`링크를 검색해 보세요.`} />
+        <input placeholder="링크를 검색해 보세요." />
       </div>
     </>
   );
